Split VRMTests into 2 files.
